+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## develop branch
 
-Nothing yet.
+### New
+
+* Filters/FilterColumns - PHP equiv of "awk { print $n }"